refactor: api요청 폴더별로 분류
