+++ frontend/src/__test__/vote/useGetVote.test.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { BrowserRouter } from 'react-router-dom';
 
-import { registerVoteItems } from '@/api/vote';
+import { registerVoteItems } from '@/api/vote/vote';
 import useVote from '@/hooks/vote/useGetVote';
 import { VoteHandler } from '@/mock';
 import { theme } from '@/styles/Theme';
