+++ frontend/src/components/@helper/accessHandler/LoginHandler.tsx
@@ -3,7 +3,7 @@ import { useEffect } from 'react';
 import { useMutation } from 'react-query';
 import { useNavigate, useSearchParams } from 'react-router-dom';
 
-import { postLogin } from '@/api/login';
+import { postLogin } from '@/api/user/login';
 import LoginLoading from '@/components/login/LoginLoading/LoginLoading';
 import { ACCESSTOKEN_KEY } from '@/constants';
 import { URL } from '@/constants/url';
