+++ frontend/src/hooks/comment/useDeleteComment.tsx
@@ -2,7 +2,7 @@ import { AxiosError } from 'axios';
 import { useEffect } from 'react';
 import { useMutation } from 'react-query';
 
-import { deleteComments } from '@/api/comments';
+import { deleteComments } from '@/api/comment/comments';
 import { ErrorMessage } from '@/constants/ErrorMessage';
 import useThrowCustomError from '@/hooks/common/useThrowCustomError';
 import { queryClient } from '@/index';
