+++ frontend/src/hooks/login/useDeleteRefreshToken.tsx
@@ -1,7 +1,7 @@
 import { AxiosError, AxiosResponse } from 'axios';
 import { useMutation } from 'react-query';
 
-import { deleteRefreshToken } from '@/api/login';
+import { deleteRefreshToken } from '@/api/user/login';
 import { ErrorMessage } from '@/constants/ErrorMessage';
 import useThrowCustomError from '@/hooks/common/useThrowCustomError';
