+++ frontend/src/hooks/login/useGetLoginURL.tsx
@@ -2,7 +2,7 @@ import { AxiosError } from 'axios';
 import { useEffect, useState } from 'react';
 import { useQuery } from 'react-query';
 
-import { getGithubURL } from '@/api/login';
+import { getGithubURL } from '@/api/user/login';
 import { ErrorMessage } from '@/constants/ErrorMessage';
 import useThrowCustomError from '@/hooks/common/useThrowCustomError';
