+++ frontend/src/hooks/tempArticle/useGetTempArticles.tsx
@@ -1,7 +1,7 @@
 import { AxiosError } from 'axios';
 import { useQuery } from 'react-query';
 
-import { getTempArticles } from '@/api/tempArticle';
+import { getTempArticles } from '@/api/tempArticle/tempArticle';
 import { ErrorMessage } from '@/constants/ErrorMessage';
 import useThrowCustomError from '@/hooks/common/useThrowCustomError';
 import { TempArticleResponse } from '@/types/articleResponse';
