+++ frontend/src/utils/takeToastImgEditor.ts
@@ -1,6 +1,6 @@
 import { MutableRefObject } from 'react';
 
-import { postImageUrlConverter } from '@/api/image';
+import { postImageUrlConverter } from '@/api/article/image';
 import { Editor } from '@toast-ui/react-editor';
 
 export const takeToastImgEditor = (content: MutableRefObject<Editor | null>) => {
